fix uploadlocation
Changed props and functions moved to another file
