+++ src/components/atoms/UploadLocation/index.js
@@ -1,104 +1,20 @@
 import React from 'react';
-import {View, TouchableOpacity, Alert, Image} from 'react-native';
-import {request, PERMISSIONS, check, RESULTS} from 'react-native-permissions';
-import RNAndroidLocationEnabler from 'react-native-android-location-enabler';
-import RNLocation from 'react-native-location';
+import {View, TouchableOpacity, Image} from 'react-native';
 import {Icon} from 'react-native-elements';
-import ModalMap from '../ModalMap';
 import styles from './styles';
+import functions from './functions';
 
-const location = (changePost, changeModalVisible) => {
-  RNLocation.configure({
-    distanceFilter: 5.0,
-    interval: 2,
-  });
-
-  const unSub = RNLocation.subscribeToLocationUpdates(async result => {
-    await changePost(
-      {
-        latitude: result[0].latitude,
-        longitude: result[0].longitude,
-        latitudeDelta: 0.09,
-        longitudeDelta: 0.04,
-      },
-      'location',
-    );
-    await changeModalVisible(true);
-    unSub();
-  });
-};
-
-const checkPermissions = (changeModalVisible, changePost) => {
-  check(PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION)
-    .then(locationresult => {
-      check(PERMISSIONS.ANDROID.WRITE_EXTERNAL_STORAGE).then(
-        externalStorage => {
-          if (
-            locationresult === RESULTS.GRANTED &&
-            externalStorage === RESULTS.GRANTED
-          ) {
-            startGPS(changePost, changeModalVisible);
-          } else {
-            requestPermissions(changePost, changeModalVisible);
-          }
-        },
-      );
-    })
-    .catch(error => Alert.alert('Error', '' + error, [{text: 'OK'}]));
-};
-
-const requestPermissions = (changePost, changeModalVisible) => {
-  request(PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION).then(locationresult2 => {
-    request(PERMISSIONS.ANDROID.WRITE_EXTERNAL_STORAGE).then(
-      externalStorage2 => {
-        if (
-          locationresult2 === RESULTS.GRANTED &&
-          externalStorage2 === RESULTS.GRANTED
-        ) {
-          startGPS(changePost, changeModalVisible);
-        } else {
-          Alert.alert(
-            'Se requieren permisos',
-            '' + 'Tienes que aceptar los permisos para usar esta función.',
-            [{text: 'OK'}],
-          );
-        }
-      },
-    );
-  });
-};
-
-const startGPS = async (changePost, changeModalVisible) => {
-  RNAndroidLocationEnabler.promptForEnableLocationIfNeeded({
-    interval: 10000,
-    fastInterval: 5000,
-  }).then(() => location(changePost, changeModalVisible));
-};
-
-export default function UploadLocation({
-  changePost,
-  setMapOpen,
-  post,
-  mapOpen,
-}) {
-  const changeModalVisible = value => setMapOpen(value);
+export default function UploadLocation({modals, handleData, dataForm}) {
   return (
     <View>
-      <ModalMap
-        visible={mapOpen}
-        changePost={changePost}
-        setMapOpen={setMapOpen}
-        init={post.valuesPost.location}
-        post={post}
-      />
       <View style={styles.containerImage}>
-        {post.valuesPost.urlMap.length > 0 ? (
-          <Image style={styles.image} source={{uri: post.valuesPost.urlMap}} />
+        {dataForm.object.urlMap.length > 0 ? (
+          <Image style={styles.image} source={{uri: dataForm.object.urlMap}} />
         ) : null}
       </View>
       <TouchableOpacity
         style={styles.btn}
-        onPress={() => checkPermissions(changeModalVisible, changePost)}>
+        onPress={() => functions.checkPermissions(modals, handleData)}>
         <Icon name={'map-pin'} type={'feather'} size={30} />
       </TouchableOpacity>
     </View>
